refactor(tests): repoint the fm-html-map test suite at the vendored copy
9 files, matching the exact require("../../scripts/lib/renderer.js")
convention already used for dsHtmlMap throughout this same suite (several
of these files, e.g. golden-snapshot and ds-screen-tree, already do both
side by side). Full suite green with the vendored path; the local copy is
now unreferenced anywhere except its own two files, retired next.
