+++ plugins/actian-design-system/tests/renderers/ds-screen-tree.test.js
@@ -380,7 +380,7 @@ test("screenTree — no header: root has no global-header child", function () {
 // flow-renderer.test.js) with dsScreenTree in the mock window.
 
 test("HTML golden — studio screen byte-identical after refactor", function () {
-  var fmHtmlMap = require("../../scripts/renderers/html-renderers/fm-html-map");
+  var fmHtmlMap = require("../../scripts/lib/renderer.js").fmHtmlMap;
   var renderNodeModule = require("../../scripts/renderers/html-renderers/render-node.js");
   var dsHtmlMap = require("../../scripts/lib/renderer.js").dsHtmlMap;
 
@@ -443,7 +443,7 @@ test("HTML golden — studio screen byte-identical after refactor", function () 
 });
 
 test("HTML golden — explorer screen byte-identical after refactor", function () {
-  var fmHtmlMap = require("../../scripts/renderers/html-renderers/fm-html-map");
+  var fmHtmlMap = require("../../scripts/lib/renderer.js").fmHtmlMap;
   var renderNodeModule = require("../../scripts/renderers/html-renderers/render-node.js");
   var dsHtmlMap = require("../../scripts/lib/renderer.js").dsHtmlMap;
   var mockDoc = {
@@ -487,7 +487,7 @@ test("HTML golden — explorer screen byte-identical after refactor", function (
 });
 
 test("HTML golden — admin screen byte-identical after refactor", function () {
-  var fmHtmlMap = require("../../scripts/renderers/html-renderers/fm-html-map");
+  var fmHtmlMap = require("../../scripts/lib/renderer.js").fmHtmlMap;
   var renderNodeModule = require("../../scripts/renderers/html-renderers/render-node.js");
   var dsHtmlMap = require("../../scripts/lib/renderer.js").dsHtmlMap;
   var mockDoc = {
@@ -528,7 +528,7 @@ test("HTML golden — admin screen byte-identical after refactor", function () {
 });
 
 test("HTML golden — no-sidebar screen byte-identical after refactor", function () {
-  var fmHtmlMap = require("../../scripts/renderers/html-renderers/fm-html-map");
+  var fmHtmlMap = require("../../scripts/lib/renderer.js").fmHtmlMap;
   var renderNodeModule = require("../../scripts/renderers/html-renderers/render-node.js");
   var dsHtmlMap = require("../../scripts/lib/renderer.js").dsHtmlMap;
   var mockDoc = {
