+++ plugins/actian-design-system/tests/renderers/flow-renderer.test.js
@@ -73,7 +73,7 @@ const mockDocument = {
 // functions. (The IIFE runs under `new Function`, whose global `require`
 // cannot resolve flow-renderer's relative paths, so we provide both deps on
 // the mocked window — mirroring how the browser bundle inlines them.)
-const fmHtmlMap = require("../../scripts/renderers/html-renderers/fm-html-map");
+const fmHtmlMap = require("../../scripts/lib/renderer.js").fmHtmlMap;
 const renderNodeModule = require("../../scripts/renderers/html-renderers/render-node.js");
 const dsHtmlMap = require("../../scripts/lib/renderer.js").dsHtmlMap;
 const dsScreenTree = require("../../scripts/renderers/html-renderers/ds-screen-tree.js");
