+++ plugins/actian-design-system/tests/renderers/fm-html-map.test.js
@@ -10,8 +10,8 @@
 
 const path = require("path");
 const { renderFMComponent, parseVariant } = require(
-  path.join(__dirname, "../../scripts/renderers/html-renderers/fm-html-map.js"),
-);
+  path.join(__dirname, "../../scripts/lib/renderer.js"),
+).fmHtmlMap;
 
 // ---------------------------------------------------------------------------
 // Minimal test harness
